requirements.txt
# Core Libraries
pymupdf
pdf2image
pytesseract
Pillow
faiss-cpu
streamlit
sentence-transformers
requests
python-dotenv
werkzeug
